Reorganized pipeline directory structure
Added option to unpack tar files for ligand and complex files. Fixed
MBAR free energy estimate. Added option to redo free energy estimates.
Moved dock6 mol2 loading to IO.py. Rudimentary smart darting.
jobs_on_queue more general.

+++ Pipeline/prep_complex_for_AlGDock.py
@@ -1,16 +1,13 @@
-# Prepares a ligand for AlGDock
-# Run from: [TARGET]/ligand/AlGDock_in/
+# Prepares a complex for AlGDock
 
 try:
   import argparse
   parser = argparse.ArgumentParser()
   parser.add_argument('ligand_mol2', default=None,
     help='Input mol2 of the ligand with sybyl atom types')
-  parser.add_argument('ligand_frcmod', default=None,
-    help='Input frcmod of the ligand')
   parser.add_argument('receptor_pdb', default=None,
     help='Input PDB of the receptor with AMBER atom types')
-  parser.add_argument('--complex_prefix', default=None,
+  parser.add_argument('complex_tarball', default=None,
     help='Prefix for the complex prmtop and inpcrd files')
   parser.add_argument('--debug', action='store_true')
   args = parser.parse_args()
@@ -18,13 +15,15 @@ except:
   import sys
   class args:
     ligand_mol2 = sys.argv[1]
-    ligand_frcmod = sys.argv[2]
-    receptor_pdb = sys.argv[3]
+    receptor_pdb = sys.argv[2]
 
 import os
-for FN in [args.ligand_mol2, args.ligand_frcmod, args.receptor_pdb]:
+for FN in [args.ligand_mol2, args.receptor_pdb]:
   if not os.path.isfile(FN):
     raise Exception('Input file %s not found!'%FN)
+args.ligand_mol2 = os.path.abspath(args.ligand_mol2)
+args.receptor_pdb = os.path.abspath(args.receptor_pdb)
+args.complex_tarball = os.path.abspath(args.complex_tarball)
 
 import os, inspect
 dirs = {'script':os.path.dirname(os.path.abspath(\
@@ -32,30 +31,39 @@ dirs = {'script':os.path.dirname(os.path.abspath(\
 execfile(os.path.join(dirs['script'],'_external_paths.py'))
 command_paths = findPaths(['sander'])
 dirs['amber'] = os.path.abspath(os.path.dirname(command_paths['sander'])[:-4])
+dirs['temp'] = args.complex_tarball + '.tmp'
 
-ligand_prefix = os.path.basename(args.ligand_mol2)[:-5]
-receptor_prefix = os.path.basename(args.receptor_pdb)[:-4]
-if args.complex_prefix is None:
-  complex_prefix = ligand_prefix + '-' + receptor_prefix
-else:
-  complex_prefix = args.complex_prefix
+if not os.path.isdir(dirs['temp']):
+  os.system('mkdir -p '+dirs['temp'])
+os.chdir(dirs['temp'])
 
-if not os.path.isdir(ligand_prefix):
-  os.makedirs(ligand_prefix)
-os.chdir(ligand_prefix)
-if not os.path.isfile(os.path.join('..',ligand_prefix+'.mol2')):
+ligand_prefix = '.'.join(os.path.dirname(args.ligand_mol2).split('/')[-1].split('.')[:-1]) \
+  + '.' + os.path.basename(args.ligand_mol2)[:-5]
+# The receptor file name ends with '.pdb2pqr_amber.pqr',
+# which is 18 characters long
+receptor_prefix = os.path.basename(args.receptor_pdb)[:-18]
+complex_prefix = ligand_prefix + '-' + receptor_prefix
+
+if not os.path.isfile(ligand_prefix+'.mol2'):
   print '\n*** Writing mol2 file with amber atom types ***'
   command = dirs['amber']+'/bin/antechamber' + \
-    ' -i {0} -fi mol2 -o {1}.mol2 -fo mol2 -rn {2}'.format(\
-      args.ligand_mol2,os.path.join('..',ligand_prefix),ligand_prefix)
+    ' -i {0} -fi mol2 -o {1}.mol2 -fo mol2 -rn LIG'.format(\
+      args.ligand_mol2,ligand_prefix)
   os.system(command)
-os.chdir('..')
-if not args.debug:
-  os.system('rm -rf '+ligand_prefix)
+  if not os.path.isfile(ligand_prefix+'.mol2'):
+    print command
+    raise Exception('Could not write mol2 file')
 
-if not (os.path.isfile(os.path.join('..',complex_prefix+'.prmtop')) and \
-        os.path.isfile(os.path.join('..',complex_prefix+'.inpcrd'))):
-  print '\n*** Generating prmtop and inpcrd files ***'
+if not os.path.isfile(ligand_prefix+'.frcmod'):
+  print '\n*** Generating frcmod file ***'
+  command = dirs['amber']+'/bin/parmchk' +\
+    ' -i {0}.mol2 -f mol2 -o {0}.frcmod -a Y -w Y'.format(ligand_prefix)
+  os.system(command)
+
+if not (os.path.isfile(os.path.join(complex_prefix+'.prmtop')) and \
+        os.path.isfile(os.path.join(complex_prefix+'.inpcrd')) and \
+        os.path.isfile(os.path.join(complex_prefix+'.pdb'))):
+  print '\n*** Generating prmtop and inpcrd and pdb files ***'
   tleap_F = open(complex_prefix+'.tleap','w')
   tleap_F.write("""
 source leaprc.ff14SB
@@ -66,28 +74,38 @@ receptor = loadpdb {0}
 
 # Ligand
 source leaprc.gaff
-loadamberparams {1}
-ligand = loadmol2 {2}.mol2
-saveoff ligand {2}.lib
-loadoff {2}.lib
+loadamberparams {1}.frcmod
+ligand = loadmol2 {1}.mol2
+saveoff ligand {1}.lib
+loadoff {1}.lib
 
 # Complex 
 complex = combine {{receptor, ligand}}
-saveamberparm complex {3}.prmtop {3}.inpcrd
+saveamberparm complex {2}.prmtop {2}.inpcrd
+savepdb complex {2}.pdb
 
 quit
-""".format(args.receptor_pdb, args.ligand_frcmod, ligand_prefix, complex_prefix))
+""".format(args.receptor_pdb, ligand_prefix, complex_prefix))
   tleap_F.close()
   command = dirs['amber']+'/bin/tleap -f {0}.tleap'.format(complex_prefix)
   os.system(command)
 
-if args.debug:
-  if os.path.isfile('leap.log'):
-    os.rename('leap.log', complex_prefix+'.leaplog')
-else:
-  for FN in [complex_prefix+'.tleap',ligand_prefix+'.mol2',ligand_prefix+'.lib', \
-    'leap.log', \
-    'ANTECHAMBER_AC.AC', 'ANTECHAMBER_AC.AC0', \
-    'ANTECHAMBER_BOND_TYPE.AC', 'ANTECHAMBER_BOND_TYPE.AC0', 'ATOMTYPE.INF']:
-    if os.path.isfile(FN):
-      os.remove(FN)
+if os.path.isfile(os.path.join(complex_prefix+'.pdb')):
+  print '\n*** Setting fixed atoms in pdb file ***'
+  command = 'python {0}/label_fixed_atoms.py {1}'
+  command = command.format(dirs['script'], os.path.join(complex_prefix+'.pdb'))
+  os.system(command)
+
+# Compresses the complex files in a tarball
+import tarfile
+tarF = tarfile.open(args.complex_tarball,'w:gz')
+tarF_contents = [complex_prefix+'.'+ext for ext in ['prmtop', 'inpcrd', 'pdb']]
+for FN in tarF_contents:
+  tarF.add(FN)
+tarF.close()
+
+os.chdir('..')
+
+if not args.debug:
+  os.system('rm -rf '+dirs['temp'])
+
